RBCoreDataManager is no longer a singleton.
Instead of having a singleton method, it has a default instance method. This way it can still be treated like a singleton, yet also be able to create new instances if necessary. This is useful for writing tests.

I have also removed some potentially unsafe and deprecated methods.

The documentation has been updated to reflect the changes.